glitched show page
